Create MatrixMinFinder.cpp
This C++ program generates a random 10x10 matrix of floating-point numbers, extracts a portion of it defined by specified indices, and then finds the minimum values in the first four columns and the last four columns of the extracted portion. Finally, it prints out the minimum values found in each segment.